+++ src/index.js
@@ -1,3 +1,3 @@
 import './components/buttons/buttons';
-import {ForceHorse} from './components/force-horse/force-horse';
+import {ForceHorse} from './components/force-horse/force-horse.component';
 export {ForceHorse};
